+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## `v1.4.0` (contracts not yet deployed; all code not yet pushed)
 
 * Added minting of additional 10 million RGT.
+* Added distributions of 750,000 RGT for the first year of liquidity mining.
+* Added vesting of 7 million RGT.
 * Added RGT distributions for the official Uniswap V2 RGT/ETH pair.
 * Added RGT distributions for RGT liquidity providers on Loopring.